Add OAuth Service as new project for DMS backend.
