+++ src/dms/backend/EdFi.DataManagementService.Backend.OAuthService/IOAuthManager.cs
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+using Microsoft.AspNetCore.Http;
+
+namespace EdFi.DmsConfigurationService.Backend.OAuthService;
+
+public interface IOAuthManager
+{
+    public Task GetAccessTokenAsync(HttpContext httpContext, string upstreamUri);
+}
